Crate aptos_cli_common

Common modules for creating an Aptos CLI command.

Modules

utils

Structs

AccountAddressWrapper

A wrapper around AccountAddress to be more flexible from strings than AccountAddress

ChangeSummary

A summary of a WriteSetChange for easy printing

EncodingOptions

An insertable option for use with encodings.

FaucetOptions

GasOptions

Gas price options for manipulating how to prioritize transactions

PrivateKeyInputOptions

ProfileOptions

PromptOptions

An insertable option for use with prompts.

PublicKeyInputOptions

RestOptions

Options specific to using the Rest endpoint

TransactionOptions

Common options for interacting with an account for a validator

TransactionSummary

A shortened transaction output

WriteTransactionOptions

Options specific to submitting a private key to the Rest endpoint

Enums

EncodingType

Types of encodings used by the blockchain

KeyType

Types of Keys used by the blockchain

Constants

DEFAULT_FAUCET_URL

DEFAULT_GAS_UNIT_PRICE

DEFAULT_MAX_GAS

DEFAULT_REST_URL

Traits

ExtractPublicKey

Functions

account_address_from_public_key

load_account_arg

Loads an account arg and allows for naming based on profiles